pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "self-evolving-agents"
version = "0.1.0"
description = "A modular framework for self-evolving LLM agents — based on survey 2507.21046 & 2508.07407"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "litellm>=1.40.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "tenacity>=8.0",
    "numpy>=1.26",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
